Use flake for dev environment
It is complicated to maintain two dev environments, both with nix and
poetry.
The switch was made to uv, since it uses the default dependency style,
it is faster and out of the box compatible with pyproject.nix.
With pyproject.nix only a pyproject.toml file needs to be provided and a
devshell will be created with all needed dependencies loaded from
nixpkgs!